fix build issue on Linux/Sparc
git-svn-id: svn://libopenr2.org/openr2/branches/release-1@304 c87312dc-f89b-4d16-81b4-35530099f826

+++ src/r2log.c
@@ -48,7 +48,7 @@ void openr2_log_channel_default(openr2_chan_t *r2chan, openr2_log_level_t level,
 	/* Avoid infinite recursion: Don't call openr2_chan_get_number 
 	   because that will call openr2_log */
 	printf("[%02d:%02d:%03lu][%s] Channel %d -- ", currtime_tm.tm_min, currtime_tm.tm_sec, 
-			currtime.tv_usec/1000, openr2_log_get_level_string(level), r2chan->number);
+			(unsigned long)currtime.tv_usec/1000, openr2_log_get_level_string(level), r2chan->number);
 	if (r2chan->r2context->configured_from_file) {
 		printf("M -- ");
 	}
@@ -81,7 +81,7 @@ static void log_at_file(openr2_chan_t *r2chan, const char *fmt, va_list ap)
 	/* Avoid infinite recurstion: Don't call openr2_chan_get_number 
 	   because that will call openr2_log */
 	fprintf(r2chan->logfile, "[%02d:%02d:%02d:%03lu] [Thread: %02lu] [Chan %d] - ", currtime_tm.tm_hour, currtime_tm.tm_min, 
-			currtime_tm.tm_sec, currtime.tv_usec/1000, (unsigned long)pthread_self(), r2chan->number);
+			currtime_tm.tm_sec, (unsigned long)currtime.tv_usec/1000, (unsigned long)pthread_self(), r2chan->number);
 	if (r2chan->r2context->configured_from_file) {
 		fprintf(r2chan->logfile, "M - ");
 	}
